invoice dashboard stats

+++ src/main/java/com/gkev/InvoicingSystem/models/DTO/InvoiceItemsResDTO.java
@@ -5,7 +5,7 @@ import java.math.BigDecimal;
 public record InvoiceItemsResDTO(
        String itemName,
        BigDecimal unitPrice,
-       double quantity,
+       BigDecimal quantity,
        BigDecimal tax,
        BigDecimal tax_total,
        BigDecimal total
